+++ Changelog.md
@@ -26,3 +26,9 @@
 - Encoding Config files to base64, base32, base16, and hex in memory
 
 ---
+
+## 1.0.4 (MAJOR)
+- Fixed file not writing to disk
+- Changed add_content to be a protected method, use update_content instead!
+- Added Decoding base64, base32, base16, and hex to Config files in memory
+---